Guardar contenido, lista del contenido, usaurio en sesión

+++ src/dao/UserDAO.java
@@ -14,110 +14,118 @@ public class UserDAO {
 
 	private static SessionFactory sf;
 	private static Session session;
-	
+
 	public static Set<User> getAllUsers() {
-		
+
 		List<User> users = null;
-		
+
 		sf = HibernateUtil.getSessionFactory();
 
 		session = sf.openSession();
 		session.beginTransaction();
-		
-		users = (List<User>)session.createQuery("FROM User").list();
-		
+
+		users = (List<User>) session.createQuery("FROM User").list();
+
 		session.getTransaction().commit();
 		session.close();
-		
+
 		Set<User> result = new HashSet<User>(users);
-		
+
 		return result;
-		
-		
+
 	}
-	
+
 	public static boolean addUser(String username, String password) {
+		
+	
 		User usr = null;
-		
+
 		sf = HibernateUtil.getSessionFactory();
-		
-		
-		
-		if (!UserDAO.validate(username, password)) {
+
+		if (!UserDAO.tryLogIn(username, password)) {
 			session = sf.openSession();
 			session.beginTransaction();
-			
+
 			usr = new User();
 			usr.setUsername(username);
 			usr.setPassword(password);
 			session.save(usr);
-			
+
 			session.getTransaction().commit();
 			session.close();
-			
+
 			return true;
-		}
-		else {
+		} else {
 			return false;
 		}
 	}
-	
+
 	public static void deleteUser(String username) {
+
+		if (username != null) {
+			
+			User use = getUser(username);
+			
+			if (use!=null) {
+				sf = HibernateUtil.getSessionFactory();
+				
+				session = sf.openSession();
+				session.beginTransaction();
+
+
+				session.delete(use);
+
+				session.getTransaction().commit();
+				session.close();
+			}
+			
+		}
 		
-		Integer num = Integer.parseInt(username);
-		sf = HibernateUtil.getSessionFactory();
-		
+	}
+
+	public static boolean tryLogIn(String username, String password) {
+
+		if (username != null && password != null) {
+			sf = HibernateUtil.getSessionFactory();
 
 			session = sf.openSession();
 			session.beginTransaction();
-			
-			User use = session.get(User.class, num);
-			
-			session.delete(use);
-			
+
+			User user = null;
+
+			user = (User) session.createQuery("FROM User U WHERE U.username = :username").setParameter("username", username)
+					.uniqueResult();
+
 			session.getTransaction().commit();
 			session.close();
+			
+			return (user != null && user.getPassword().equals(password));
 		}
-	
-	public static boolean validate(String username, String password) {
 		
-		sf = HibernateUtil.getSessionFactory();
+		return false;
 
-		session = sf.openSession();
-		session.beginTransaction();
-		
-		User user = null;
-		
-		user = (User) session.createQuery("FROM User U WHERE U.username = :username").setParameter("username", username)
-                .uniqueResult();
-		
-		session.getTransaction().commit();
-		session.close();
-		
-		return (user != null && user.getPassword().equals(password));
-		
-		
-		
 	}
-	
-public static boolean validate(String username) {
+
+	public static boolean validate(String username) {
+		
+		return getUser(username) != null;
+	}
+
+	public static User getUser(String username) {
 		
 		sf = HibernateUtil.getSessionFactory();
 
 		session = sf.openSession();
 		session.beginTransaction();
-		
+
 		User user = null;
-		
+
 		user = (User) session.createQuery("FROM User U WHERE U.username = :username").setParameter("username", username)
-                .uniqueResult();
-		
+				.uniqueResult();
+
 		session.getTransaction().commit();
 		session.close();
-		
-		return (user != null );
-		
-		
-		
+
+		return user;
 	}
 }
